Address Jaliya feedback
