+++ torch/utils/data/__init__.pyi
@@ -1,6 +1,6 @@
 from .sampler import Sampler as Sampler, SequentialSampler as SequentialSampler, RandomSampler as RandomSampler, \
     SubsetRandomSampler as SubsetRandomSampler, WeightedRandomSampler as WeightedRandomSampler, BatchSampler as BatchSampler
-from .distributed import DistributedSampler as DistributedSampler
+from .distributed import DistributedSampler as DistributedSampler, DistributedChunkSampler as DistributedChunkSampler
 from .dataset import Dataset as Dataset, TensorDataset as TensorDataset, ConcatDataset as ConcatDataset, \
     Subset as Subset, random_split as random_split
 from .dataloader import DataLoader as DataLoader
