+++ torch/utils/data/dataset.pyi
@@ -29,10 +29,6 @@ class IterableDataset(Dataset[T_co]):
     def __iter__(self) -> None: ...
     def __add__(self, other: Any): ...
 
-class ChunkDataReader:
-    def __init__(self) -> None: ...
-    def __call__(self, idx: Any) -> None: ...
-
 class ChunkDataset(IterableDataset):
     chunk_sampler: Any = ...
     chunk_reader: Any = ...
