+++ torch/utils/data/sampler.pyi
@@ -22,14 +22,3 @@ class WeightedRandomSampler(Sampler[int]):
 
 class BatchSampler(Sampler[List[int]]):
     def __init__(self, sampler: Sampler[int], batch_size: int, drop_last: bool) -> None: ...
-
-class DistributedChunkSampler(Sampler[int]):
-    world_size: Any = ...
-    rank: Any = ...
-    num_chunks: Any = ...
-    shuffle: Any = ...
-    sampler: Any = ...
-    def __init__(self, world_size: Any, rank: Any, num_chunks: Any, shuffle: Any, sampler: Any) -> None: ...
-    def __iter__(self) -> None: ...
-    def __len__(self) -> None: ...
-    def set_rank(self, rank: Any) -> None: ...
